unhanced the auth system for todo

+++ backend/Controllers/todoControllers.js
@@ -1,7 +1,7 @@
 const Todo = require("../Model/todoSchema");
 const auth = require("../middleware/auth");
 const { default: mongoose, Mongoose } = require("mongoose");
-const cookieParser = require('cookie-parser');
+
 
 
 //for home route
@@ -81,12 +81,18 @@ exports.getAllTodos= async (req, res) => {
 
 exports.getATodo = async (req, res) =>{
     try {
+        const uID = req.user.id;
+        console.log(uID);
+        const uniqueUser = await Todo.find({userID : uID});
+
+        if ( uniqueUser) {
         const todoID = await Todo.findById(req.params._id);
         console.log(todoID);
         res.status(200).json({
             success : true,
             todoID
         })
+    }
     } catch (error) {
         console.log(error);
         console.log("Failed to get a todo!");
@@ -100,13 +106,23 @@ exports.getATodo = async (req, res) =>{
 //edit todo title
 exports.editTodoTitle = async (req, res) => {
     try {
+        const uID = req.user.id;
+        console.log(uID);
+        const uniqueUser = await Todo.find({userID : uID});
+
+        if (uniqueUser) {
+            
+        
+            
         const editedTodo = await Todo.findByIdAndUpdate(req.params._id, req.body);
+                console.log(req.body);
             console.log(editedTodo)
+           const savedTodo =  await editedTodo.save();
+           console.log(savedTodo)
         res.status(201).json({
             success : true,
             message :"Todo edited!",
-         
-        })
+        })}
     } catch (error) {
         console.log(error);
         console.log("Failed to edit a todo!");
@@ -121,12 +137,18 @@ exports.editTodoTitle = async (req, res) => {
 //delete todo 
 exports.deleteTodo= async (req, res) => {
     try {
+
+        const uID = req.user.id;
+        console.log(uID);
+        const uniqueUser = await Todo.find({userID : uID});
+
+        if (uniqueUser) {
         const deletedTodo = await Todo.findByIdAndDelete(req.params._id);
             console.log(deletedTodo)
         res.status(201).json({
             success : true,
             message :"Todo deleted!"
-        })
+        })}
     } catch (error) {
         console.log(error);
         console.log("Failed to delete a todo!");
@@ -142,19 +164,28 @@ exports.deleteTodo= async (req, res) => {
 exports.tasksForATodo = async (req, res) =>{
 
     try {
+
+        const uID = req.user.id;
+        console.log(uID);
+        const uniqueUser = await Todo.find({userID : uID});
+
+        if (uniqueUser) {
+
         const todo = await Todo.findById(req.params._id);
 
         if (!todo) {
             return res.status(401).send("Todo doesn't exist!");}
     
             const taskToBeAdded = req.body.Tasks; //It'll acces the key of an Object and will return the value
+                //use "Tasks" as key in postman
             todo.Tasks.push(taskToBeAdded);
             await todo.save();
+            console.log(todo);
             res.status(201).json({
                 success : true,
                 message : "Task added in your todo!",
                 todo 
-            })
+            })}
     } catch (error) {
         console.log(error);
         console.log("Failed to add a task!");
@@ -169,6 +200,13 @@ exports.tasksForATodo = async (req, res) =>{
 exports.getAllTasksForATodo = async (req, res) =>{
 
     try {
+
+        const uID = req.user.id;
+        console.log(uID);
+        const uniqueUser = await Todo.find({userID : uID});
+
+        if (uniqueUser) {
+
         const todo = await Todo.findById(req.params._id);
 
 if (!todo) { res.status(401).send("Todo doesn't exisit");
@@ -181,7 +219,7 @@ console.log( todo);
     res.status(201).json({
         success  : true,
         allTasks 
-    })
+    })}
     } catch (error) {
         console.log(error);
         console.log("Failed to show all tasks!");
@@ -195,6 +233,12 @@ console.log( todo);
 //edit a task for a specific todo
 exports.editTaskForATodo = async (req, res) => {
     try {
+
+        const uID = req.user.id;
+        console.log(uID);
+        const uniqueUser = await Todo.find({userID : uID});
+
+        if (uniqueUser) {
         const targetTodo = await Todo.findById(req.params._id);
         if (!targetTodo) {
             return res.status(401).send("Required todo doesn't exist");
@@ -202,13 +246,14 @@ exports.editTaskForATodo = async (req, res) => {
 
         //type key in the postman word to word
         const {taskIndex,newTaskText} = req.body;
+       
         targetTodo.Tasks.splice(taskIndex,1,newTaskText);
         await targetTodo.save();
         res.status(200).json({
             success : true,
             message : "Task edited!",
             targetTodo
-        })
+        })}
     } catch (error) {
         console.log(error);
         console.log("Failed to edit task for required todo!");
@@ -222,6 +267,11 @@ exports.editTaskForATodo = async (req, res) => {
 //delete a task for a specific todo
 exports.deleteTaskForATodo = async (req, res) => {
     try {
+        const uID = req.user.id;
+        console.log(uID);
+        const uniqueUser = await Todo.find({userID : uID});
+
+        if (uniqueUser) {
         const targetTodo = await Todo.findById(req.params._id);
     if (!targetTodo) {
         return res.status(401).send("Required todo doesn't exist");
@@ -235,7 +285,7 @@ exports.deleteTaskForATodo = async (req, res) => {
         success :true,
         message :"Task got deleted",
         targetTodo
-    })
+    })}
 
     } catch (error) {
         console.log(error);
@@ -251,20 +301,32 @@ exports.deleteTaskForATodo = async (req, res) => {
 exports.toSearch = async (req, res) =>{
 
     try {
+
+        const uID = req.user.id;
+        console.log(uID);
+        const uniqueUser = await Todo.find({userID : uID});
+
+        if (uniqueUser) {
         const {search} = req.query;
-
-
         console.log("-->",search);
         if (!search) {
             res.status(401).send("Please enter text to search!")
         }
     
-        const searchedTodos = await Todo.find({ $or : [{"Title": {$regex: search, $options:'i'}},{"Tasks": {$regex: search, $options:'i'}}]})
-
-        res.status(200).json({
-            success :true,
-            searchedTodos
-        })
+        const searchedTodos = await Todo.find(  {  $and :[ {userID : uID},{$or : [{"Title": {$regex: search, $options:'i'}},{"Tasks": {$regex: search, $options:'i'}}]}]})
+                console.log(searchedTodos);
+                if (searchedTodos.length>0) {
+                    res.status(200).json({
+                        success :true,
+                        searchedTodos
+                    })
+                } else {
+                    res.status(200).json({
+                        success :false,
+                        message : "No such todo or task exist!"
+                    })
+                }
+       }
 
     } catch (error) {
         console.log(error);
@@ -283,10 +345,15 @@ exports.toSearch = async (req, res) =>{
 //to Sort according to date and time
 
 exports.sortDateAndTime = async (req, res) =>{
-    const sortedTodosAtCreation = await Todo.find().sort({createdAt: -1});
-    const sortedTodosAtUpdation = await Todo.find().sort({updatedAt: -1});
+    const uID = req.user.id;
+    console.log(uID);
+    const uniqueUser = await Todo.find({userID : uID});
+
+    if (uniqueUser) {
+    const sortedTodosAtCreation = await Todo.find({userID : uID}).sort({createdAt: -1});
+    const sortedTodosAtUpdation = await Todo.find({userID : uID}).sort({updatedAt: -1});
     res.status(200).json({
         sortedTodosAtCreation,
         sortedTodosAtUpdation
-    })
+    })}
 }
